refactor: exports consts and moves the files around
