+++ i18n/en.pot
@@ -613,6 +613,18 @@ msgstr ""
 msgid "You dont have access to create an event in the current selections"
 msgstr ""
 
+msgid "Review duplicates"
+msgstr ""
+
+msgid "Create {{trackedEntityTypeName}} and link"
+msgstr ""
+
+msgid "Metadata error. see log for details"
+msgstr ""
+
+msgid "Review Duplicates"
+msgstr ""
+
 msgid "A duplicate exists (but there were some errors, see log for details"
 msgstr ""
 
@@ -638,15 +650,6 @@ msgstr ""
 msgid "Link"
 msgstr ""
 
-msgid "Metadata error. see log for details"
-msgstr ""
-
-msgid "Review duplicates"
-msgstr ""
-
-msgid "Create {{trackedEntityTypeName}} and link"
-msgstr ""
-
 msgid "Possible duplicates found"
 msgstr ""
 
@@ -666,9 +669,6 @@ msgid "Organisation Unit"
 msgstr ""
 
 msgid "Registration"
-msgstr ""
-
-msgid "Review Duplicates"
 msgstr ""
 
 msgid "Search results"
