+++ src/core_modules/capture-core/components/DataEntries/Enrollment/EnrollmentDataEntry.component.js
@@ -272,13 +272,14 @@ class FinalEnrollmentDataEntry extends React.Component<FinalTeiDataEntryProps> {
         },
     };
 
+
     render() {
         const { enrollmentMetadata, programId, ...passOnProps } = this.props;
         return (
             // $FlowFixMe[cannot-spread-inexact] automated comment
             <DataEntry
-                dataEntrySections={FinalEnrollmentDataEntry.dataEntrySectionDefinitions}
                 {...passOnProps}
+                dataEntrySections={FinalEnrollmentDataEntry.dataEntrySectionDefinitions}
                 formFoundation={enrollmentMetadata.enrollmentForm}
             />
         );
@@ -310,7 +311,7 @@ class PreEnrollmentDataEntryPure extends React.PureComponent<Object> {
     }
 }
 
-class PreEnrollmentDataEntry extends React.Component<PreEnrollmentDataEntryProps> {
+export class EnrollmentDataEntryComponent extends React.Component<PreEnrollmentDataEntryProps> {
     getValidationContext = () => {
         const { orgUnit, onGetUnsavedAttributeValues } = this.props;
         return {
@@ -341,7 +342,8 @@ class PreEnrollmentDataEntry extends React.Component<PreEnrollmentDataEntryProps
             onUpdateDataEntryField,
             onStartAsyncUpdateField,
             onGetUnsavedAttributeValues,
-            ...passOnProps } = this.props;
+            ...passOnProps
+        } = this.props;
         return (
             // $FlowFixMe[cannot-spread-inexact] automated comment
             <PreEnrollmentDataEntryPure
@@ -354,5 +356,3 @@ class PreEnrollmentDataEntry extends React.Component<PreEnrollmentDataEntryProps
         );
     }
 }
-
-export default PreEnrollmentDataEntry;
