+++ src/core_modules/capture-core/components/Pages/New/NewPage.container.js
@@ -17,7 +17,6 @@ export const NewPage: ComponentType<{||}> = () => {
         () => { dispatch(showDefaultViewOnNewPage()); },
         [dispatch]);
 
-
     const error: boolean =
         useSelector(({ activePage }) => activePage.selectionsError && activePage.selectionsError.error);
