+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/DataEntry/Enrollment/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { DataEntryEnrollment } from './DataEntryEnrollment.container';
