+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/DataEntry/TrackedEntityInstance/DataEntryTrackedEntityInstance.component.js
@@ -1,9 +1,9 @@
 // @flow
 import React, { Component } from 'react';
 import { withTheme } from '@material-ui/core/styles';
-import type { RenderFoundation, TeiRegistration } from '../../../../../metaData';
+import type { RenderFoundation, TeiRegistration } from '../../../../../../metaData';
 import ConfiguredTei from './ConfiguredTei.component';
-import { DATA_ENTRY_ID } from '../registerTei.const';
+import { DATA_ENTRY_ID } from '../../registerTei.const';
 import teiClasses from './trackedEntityInstance.module.css';
 
 type Props = {
@@ -39,15 +39,13 @@ class RelationshipTrackedEntityInstance extends Component<Props> {
             ...passOnProps
         } = this.props;
         return (
-            <div>
-                {/* $FlowFixMe[cannot-spread-inexact] automated comment */}
-                <ConfiguredTei
-                    id={DATA_ENTRY_ID}
-                    onSave={this.handleSave}
-                    fieldOptions={this.fieldOptions}
-                    {...passOnProps}
-                />
-            </div>
+            // $FlowFixMe[cannot-spread-inexact] automated comment
+            <ConfiguredTei
+                id={DATA_ENTRY_ID}
+                onSave={this.handleSave}
+                fieldOptions={this.fieldOptions}
+                {...passOnProps}
+            />
         );
     }
 }
