+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/DataEntry/TrackedEntityInstance/DataEntryTrackedEntityInstance.container.js
@@ -1,9 +1,7 @@
 // @flow
 import { connect } from 'react-redux';
-import {
-    makeTeiRegistrationMetadataSelector,
-} from './tei.selectors';
-import TrackedEntityInstance from './TrackedEntityInstance.component';
+import { makeTeiRegistrationMetadataSelector } from './tei.selectors';
+import TrackedEntityInstance from './DataEntryTrackedEntityInstance.component';
 
 const makeMapStateToProps = () => {
     const teiRegistrationMetadataSelector = makeTeiRegistrationMetadataSelector();
@@ -20,6 +18,5 @@ const makeMapStateToProps = () => {
     return mapStateToProps;
 };
 
-// $FlowSuppress
 // $FlowFixMe[missing-annot] automated comment
-export default connect(makeMapStateToProps, () => ({}))(TrackedEntityInstance);
+export const DataEntryTrackedEntityInstance = connect(makeMapStateToProps, () => ({}))(TrackedEntityInstance);
