+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/DataEntry/TrackedEntityInstance/tei.selectors.js
@@ -2,8 +2,8 @@
 import log from 'loglevel';
 import { errorCreator } from 'capture-core-utils';
 import { createSelector } from 'reselect';
-import type { TrackedEntityType } from '../../../../../metaData';
-import { getTrackedEntityTypeThrowIfNotFound } from '../../../../../metaData';
+import type { TrackedEntityType } from '../../../../../../metaData';
+import { getTrackedEntityTypeThrowIfNotFound } from '../../../../../../metaData';
 
 const trackedEntityTypeIdSelector = state => state.newRelationship.selectedRelationshipType.to.trackedEntityTypeId;
