--- src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/Enrollment/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default as Enrollment } from './Enrollment.container';
