--- src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/TrackedEntityInstance/index.js
@@ -1,2 +1,0 @@
-// @flow
-export { default as TrackedEntityInstance } from './TrackedEntityInstance.container';
