+++ src/core_modules/capture-core/components/Pages/NewRelationship/RegisterTei/index.js
@@ -1,10 +1,10 @@
 // @flow
 export { RegisterTei } from './RegisterTei.container';
 export { openNewRelationshipRegisterTeiEpic } from './open.epics';
-export { openNewRelationshipRegisterTeiDataEntryEpic } from './DataEntry/dataEntry.epics';
+export { openNewRelationshipRegisterTeiDataEntryEpic } from './DataEntry/RegisterTeiDataEntry.epics';
 export { actionTypes } from './registerTei.actions';
 export { actionTypes as registrationSectionActionTypes } from './RegistrationSection';
-export { actionTypes as dataEntryActionTypes } from './DataEntry/dataEntry.actions';
+export { actionTypes as dataEntryActionTypes } from './DataEntry/RegisterTeiDataEntry.actions';
 export {
     actionTypes as searchGroupDuplicateActionTypes,
 } from './GeneralOutput/WarningsSection/SearchGroupDuplicate/searchGroupDuplicate.actions';
